+++ build.rs
@@ -1,4 +1,5 @@
-use std::{env, fs, process::Command};
+use std::process::Command;
+use std::{env, fs};
 
 fn emit_version() {
     let package_name = env::var("CARGO_PKG_NAME").unwrap();
@@ -26,5 +27,6 @@ fn main() {
     emit_version();
 
     // trigger recompilation when a new migration is added
-    println!("cargo:rerun-if-changed=migrations");
+    println!("cargo:rerun-if-changed=migrations/sqlite");
+    println!("cargo:rerun-if-changed=migrations/postgres");
 }