+++ clients/python/src/taskbroker_client/worker/__init__.py
@@ -1,3 +1,3 @@
-from .worker import TaskWorker
+from .worker import PushTaskWorker, TaskWorker
 
-__all__ = ("TaskWorker",)
+__all__ = ("TaskWorker", "PushTaskWorker")