+++ clients/CHANGELOG.md
@@ -1,4 +1,26 @@
 # Changelog
+## 0.1.11
+
+### New Features ✨
+
+- Add volatile schedule storage by @markstory in [#605](https://github.com/getsentry/taskbroker/pull/605)
+
+## 0.1.9
+
+### New Features ✨
+
+- (taskbroker) Dual-write new parameters_bytes by @untitaker in [#602](https://github.com/getsentry/taskbroker/pull/602)
+
+### Other
+
+- release: 26.5.0 by @untitaker in [0206415b](https://github.com/getsentry/taskbroker/commit/0206415bffe6b91c4934fe3e3682e78f2659f13c)
+
+## 26.5.0
+
+### New Features ✨
+
+- (taskbroker) Dual-write new parameters_bytes by @untitaker in [#602](https://github.com/getsentry/taskbroker/pull/602)
+
 ## 0.1.8
 
 ### New Features ✨